env: Fix 32Bit casting env offset located in RAM
It is wrong to cast to uint32_t as the library also supports
64 Bit execution environment for example to test on a 64 bit
host machine.

Issue: #MCOM03SW-2603

+++ libs/env/env-io-ram.c
@@ -76,7 +76,7 @@ int32_t env_io_ram_init(env_io_config_t *cfg, signed long offset, size_t size)
 	if (offset < 0)
 		return -EFORBIDDEN;
 	else
-		cfg->offset = (uint32_t)offset;
+		cfg->offset = offset;
 
 	cfg->size = size;
